Compare subnet length as well when deleting DHCP entry
When searching for the DHCP subnet entry to delete, the existing code
compares only the subnet prefix and ignores subnet length.  This could
delete the wrong entry/entries if nested subnets are present.

Closes-Bug: #1362416

+++ neutron/plugins/midonet/midonet_lib.py
@@ -212,7 +212,8 @@ class MidoClient:
             raise MidonetApiException(
                 msg=_("Tried to delete non-existent DHCP"))
         for dhcp in dhcp_subnets:
-            if dhcp.get_subnet_prefix() == net_addr:
+            if (dhcp.get_subnet_prefix() == net_addr and
+                dhcp.get_subnet_length() == str(net_len)):
                 dhcp.delete()
                 break
